Implemented hotel reservations administration.

+++ Web/TravelGuide.Web/Areas/Administration/Controllers/HotelReservationsController.cs
@@ -4,19 +4,24 @@
     using System.Linq;
     using System.Threading.Tasks;
 
+    using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.EntityFrameworkCore;
 
     using TravelGuide.Data.Common.Repositories;
     using TravelGuide.Data.Models;
     using TravelGuide.Services.Data.ServiceInterfaces;
+    using TravelGuide.Web.Controllers;
     using TravelGuide.Web.ViewModels.Administration.HotelReservations;
 
     using static TravelGuide.Common.ErrorMessages.ReservationErrorMessages;
+    using static TravelGuide.Common.GlobalConstants;
     using static TravelGuide.Common.GlobalConstants.ToastrMessageConstants;
     using static TravelGuide.Common.SuccessMessages.ReservationSuccessMessages;
 
-    public class HotelReservationsController : AdministrationController
+    [Authorize(Roles = AdministratorOrHotelier)]
+    [Area("Administration")]
+    public class HotelReservationsController : BaseController
     {
         private readonly IDeletableEntityRepository<HotelReservation> hotelReservationRepository;
         private readonly IReservationService reservationService;
@@ -31,105 +36,25 @@
 
         public async Task<IActionResult> Index()
         {
-            var hotels = await this.hotelReservationRepository.AllAsNoTracking().ToListAsync();
+            var hotelReservations = await this.reservationService.GetAllHotelReservationsAsync<HotelReservationViewModel>();
 
-            var model = new IndexViewModel()
-            {
-                Hotels = await this.reservationService.GetAllHotelReservationsAsync<HotelReservationViewModel>(),
-            };
-
-            return this.View(model);
+            return this.View(hotelReservations);
         }
 
         public async Task<IActionResult> Details(string id)
         {
             if (id == null)
             {
-                return this.NotFound();
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Index));
             }
 
-            var hotelReservation = await this.hotelReservationRepository.AllAsNoTracking().FirstOrDefaultAsync(x => x.Id.ToString() == id);
+            var hotelReservation = await this.reservationService.GetHotelReservationByIdAsync<HotelReservationViewModel>(id);
 
             if (hotelReservation == null)
             {
-                return this.NotFound();
-            }
-
-            return this.View(hotelReservation);
-        }
-
-        public IActionResult Create()
-        {
-            return this.View();
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Create(HotelReservationViewModel model)
-        {
-            if (this.ModelState.IsValid)
-            {
-                return this.View(model);
-            }
-
-            try
-            {
-                await this.reservationService.AddHotelReservationAsync(model);
-            }
-            catch (Exception ex)
-            {
-                this.TempData[ErrorMessage] = ex.Message;
-
-                return this.View(model);
-            }
-
-            this.TempData[SuccessMessage] = SuccessfullyCreatedReservation;
-
-            return this.RedirectToAction(nameof(this.Index));
-        }
-
-        public async Task<IActionResult> Edit(string id)
-        {
-            if (id == null)
-            {
-                return this.NotFound();
-            }
-
-            var hotelReservation = await this.hotelReservationRepository.AllAsNoTracking().FirstOrDefaultAsync(x => x.Id.ToString() == id);
-
-            if (hotelReservation == null)
-            {
-                return this.NotFound();
-            }
-
-            return this.View(hotelReservation);
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Edit(string id, [Bind("Id,Price,UserId,HotelId,DiscountId,StartDay,EndDay,CreatedOn,ModifiedOn,IsDeleted,DeletedOn")] HotelReservation hotelReservation)
-        {
-            if (id != hotelReservation.Id.ToString())
-            {
-                return this.NotFound();
-            }
-
-            if (this.ModelState.IsValid)
-            {
-                try
-                {
-                    this.hotelReservationRepository.Update(hotelReservation);
-                    await this.hotelReservationRepository.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!await this.HotelReservationExists(id))
-                    {
-                        return this.NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
+                this.TempData[ErrorMessage] = SomethingWentWrong;
 
                 return this.RedirectToAction(nameof(this.Index));
             }
@@ -137,18 +62,88 @@
             return this.View(hotelReservation);
         }
 
+        public async Task<IActionResult> Edit(string id)
+        {
+            if (id == null)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Index));
+            }
+
+            var hotelReservation = await this.reservationService.GetHotelReservationByIdAsync<HotelReservationViewModel>(id);
+
+            if (hotelReservation == null)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Index));
+            }
+
+            return this.View(hotelReservation);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> Edit(string id, HotelReservationViewModel model)
+        {
+            if (id == null)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Edit));
+            }
+
+            var hotelReservation = await this.reservationService.GetHotelReservationByIdAsync(id);
+
+            if (hotelReservation == null)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Edit));
+            }
+
+            if (!this.ModelState.IsValid)
+            {
+                return this.View(hotelReservation);
+            }
+
+            try
+            {
+                hotelReservation.Price = model.Price;
+                hotelReservation.StartDay = model.StartDay;
+                hotelReservation.EndDay = model.EndDay;
+
+                this.hotelReservationRepository.Update(hotelReservation);
+                await this.hotelReservationRepository.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Edit));
+            }
+
+            this.TempData[SuccessMessage] = SuccessfullyEditedReservation;
+
+            return this.RedirectToAction(nameof(this.Index));
+        }
+
         public async Task<IActionResult> Delete(string id)
         {
             if (id == null)
             {
-                return this.NotFound();
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Index));
             }
 
-            var hotelReservation = await this.hotelReservationRepository.AllAsNoTracking().FirstOrDefaultAsync(x => x.Id.ToString() == id);
+            var hotelReservation = await this.reservationService.GetHotelReservationByIdAsync<HotelReservationViewModel>(id);
 
             if (hotelReservation == null)
             {
-                return this.NotFound();
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Index));
             }
 
             return this.View(hotelReservation);
@@ -158,22 +153,37 @@
         [ActionName("Delete")]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
-            var hotelReservation = await this.hotelReservationRepository.AllAsNoTracking().FirstOrDefaultAsync(x => x.Id.ToString() == id);
-
-            this.hotelReservationRepository.Delete(hotelReservation);
-            await this.hotelReservationRepository.SaveChangesAsync();
-
-            return this.RedirectToAction(nameof(this.Index));
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
+            if (id == null)
             {
-                this.hotelReservationRepository.Dispose();
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Edit));
             }
 
-            base.Dispose(disposing);
+            var hotelReservation = await this.reservationService.GetHotelReservationByIdAsync(id);
+
+            if (hotelReservation == null)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.RedirectToAction(nameof(this.Edit));
+            }
+
+            try
+            {
+                this.hotelReservationRepository.Delete(hotelReservation);
+                await this.hotelReservationRepository.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                this.TempData[ErrorMessage] = SomethingWentWrong;
+
+                return this.NotFound();
+            }
+
+            this.TempData[SuccessMessage] = SuccessfullyDeletedReservation;
+
+            return this.RedirectToAction(nameof(this.Index));
         }
 
         private async Task<bool> HotelReservationExists(string id) => await this.hotelReservationRepository.AllAsNoTracking().AnyAsync(x => x.Id.ToString() == id);
